Apply suggestions from code review

+++ app/modules/cart/models.py
@@ -10,6 +10,6 @@ class CartItem(Base):
     product_id = Column(Integer, ForeignKey("tbl_products.id"), primary_key=True, index=True)
     quantity = Column(Integer, default=1, nullable=False)
     price = Column(Float, default=0.0, nullable=False)
-    total = Column(Float, default=0.0, nullable=False)
+    total = Column(Integer, default=0, nullable=False)
     date_created = Column(DateTime, default=func.now(), nullable=False)
     date_modified = Column(DateTime, default=func.now(), onupdate=func.now(), nullable=False)
